allow the usage of regex and better docu

+++ backup_all_docker_volumes.sh
@@ -6,19 +6,22 @@ function usage()
 {
    cat << HEREDOC
 
-   Usage: $progname [--dry-run] -d <days> -b <backupfolder>
+   Usage: backup_all_docker_volumes.sh [--dry-run] [-r <regex>] -b <backupfolder> -d <days>
 
    arguments:
      -h, --help           show this help message and exit
+     -r                   regex of which containers should be included in the backup (default: .*)
      -b                   backup location of the docker volumes
-     -d                   delete all .tar files older than d in the backup folder
      --dry-run            do a dry run, dont change any files
+     -d                   delete all .tar files older than d in the backup folder
 
    Backup all Docker volumes for containers which are currently running.
    The volumes will be attached readonly to an temporary busybox container
    which archives all contents in a .tar file in the backup folder.
 
-   Additional it deletes older backup based on their creation time.
+   All volume names containing a "/" are ommited (e.g. /var/run/docker.sock).
+
+   Additional it deletes older .tar files based on their creation date.
 
    The final folder structure of the backup folder will look like this:
         backupfolder
@@ -36,6 +39,7 @@ HEREDOC
 exit 1
 }
 
+REGEX=".*"
 BASEFOLDER=
 DRYRUN=FALSE
 DELETEDAYS=
@@ -60,6 +64,11 @@ case $key in
     shift # past argument
     shift # past value
     ;;
+    -r)
+    REGEX="$2"
+    shift # past argument
+    shift # past value
+    ;;
     *)    # unknown option
     usage
     shift # past argument
@@ -81,32 +90,36 @@ fi
 # loop over all lines of output
 docker container ls --format "{{.Names}};{{.Mounts}}" --no-trunc | while read LINE ; do
 
-    CONTAINER=$(echo "$LINE" | cut -d';' -f1)
-    VOLUMES=$(echo "$LINE" | cut -d';' -f2)
+    CONTAINER=$(echo "$LINE" | cut -d';' -f1) # container name
+    VOLUMES=$(echo "$LINE" | cut -d';' -f2) # comma separated list of volumes
 
-    # loop over all volumes
-    for VOLUME in $(echo "$VOLUMES" | sed "s/,/ /g")
-    do
+    if [[ "$CONTAINER" =~ $REGEX ]]
+    then
 
-        # avoid non valid volumes like "/var/run/docker.sock"
-        if [[ ! "$VOLUME" =~ "/" ]]
-        then
+        # loop over all volumes
+        for VOLUME in $(echo "$VOLUMES" | sed "s/,/ /g")
+        do
 
-            FOLDER="$BASEFOLDER"/"$CONTAINER"
-
-            mkdir "$FOLDER" -p
-
-            DATE=$(date +"%Y-%m-%d_%H-%M-%S")
-
-            echo "Backup $VOLUME of $CONTAINER to $FOLDER at $DATE"
-
-            if [[ $DRYRUN == "FALSE" ]]
+            # avoid non valid volumes like "/var/run/docker.sock"
+            if [[ ! "$VOLUME" =~ "/" ]]
             then
-                # Based on: https://docs.docker.com/storage/volumes/#backup-a-container
-                docker run --rm -v "$VOLUME":/volume:ro -v "$FOLDER":/backup busybox tar -czf /backup/"$VOLUME"_"$DATE".tar -C /volume .
+
+                FOLDER="$BASEFOLDER"/"$CONTAINER"
+
+                mkdir "$FOLDER" -p
+
+                DATE=$(date +"%Y-%m-%d_%H-%M-%S")
+
+                echo "Backup $VOLUME of $CONTAINER to $FOLDER at $DATE"
+
+                if [[ $DRYRUN == "FALSE" ]]
+                then
+                    # do the actual backup. Based on: https://docs.docker.com/storage/volumes/#backup-a-container
+                    docker run --rm -v "$VOLUME":/volume:ro -v "$FOLDER":/backup busybox tar -czf /backup/"$VOLUME"_"$DATE".tar -C /volume .
+                fi
             fi
-        fi
-    done
+        done
+    fi
 done
 
 # delete old folders
